refactor(compiler): Remove attribute "tile-sizes" from operations after tiling
The tiling infrastructure preserves attributes of tiled
`linalg.generic` operations, such that the attribute for the tile
sizes specified for the `linalg.generic` operation before tiling is
copied to the `linalg.generic` operation that is part of the generated
IR for a single tile.

This change causes the attribute to be removed after tiling, since it
does not make sense to preserve the attribute for per-tile operations.