perf(ci): avoid building cuda image everytime
from the docs: The jobs.<job_id>.if condition is evaluated before jobs.<job_id>.strategy.matrix is applied. So we can't just use matrix.dockerfile so we have to build both images if one of the two files change, or we will have to split this into two.

it was previously built everytime due to the evaluation of
matrix.dockerfile to an empty string

https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#jobsjob_idif