feat(compiler/optimizer): Add parser support for domains of GLWE variables
This adds support for a new attribute `glwe.domain` that allows for
the specification of all the possible values a variable used in GLWE
expressions can take. The values can either be integer or floating
point literals.

For example, the following attribute defines that the variable `n` can
take the integer values 1, 2, 3 or 4:

  #ndom = #glwe.domain<@n in [1, 2, 3, 4]>

If any of the values can only parsed as a floating point value, all
values are considered to be floating point values. E.g.,

  #ndom = #glwe.domain<@n in [1.0, 2, 3, 4]>

is equivalent to:

  #ndom = #glwe.domain<@n in [1.0, 2.0, 3.0, 4.0]>

The value of the domain attribute is a
`std::variant<SmallVector<int64_t, double>>`.