refactor(compiler/optimizer): Add parser support for global GLWE constraints
This adds a new attribute `glwe.constraint` that can be used to define
constraints on GLWE expressions. The generic form is:

  glwe.constraint<LHS CMP RHS>

where LHS and RHS are arbitrary GLWE expressions and CMP is a
comparison operator. The operator must be either (`lt` (lower than),
`le` (lower or equal), `eq` (equal), `gt` (greater than) or `ge`
(greater or equal).