feature(compiler/optimizer): Add walk function for ASTs of GLWE expressions
This adds a function template `GLWEExpr::walk()`, which can be used to
iterate over the nodes of the AST rooted at the expression. The
semantics are similar to `Operation::walk()`, including the order of
the walk and the semantics defined by the argument and return type of
the callback function.